[refactor🔨] : 전체조회 누락부분추가
